Add tooltip reading from Catalog Feature
