+++ .prettierrc
@@ -1,0 +1,4 @@
+{
+  "trailingComma": "none",
+  "arrowParens": "avoid"
+}
