+++ buildprocess/configureWebpack.js
@@ -87,6 +87,7 @@ function configureWebpack(terriaJSBasePath, config, devMode, hot, MiniCssExtract
                 'babel-plugin-jsx-control-statements',
                 '@babel/plugin-transform-modules-commonjs',
                 '@babel/plugin-syntax-dynamic-import',
+                ["@babel/plugin-proposal-optional-chaining"],
                 ['@babel/plugin-proposal-class-properties',
                     {
                         loose: true
