+++ lib/Models/Terria.js
@@ -42,6 +42,7 @@ var ViewerMode = require("./ViewerMode");
 var findIndex = require("../Core/findIndex.js");
 var i18next = require("i18next").default;
 var overrides = require("../Overrides/defaults.jsx").overrides;
+// var Receipt = require("./Receipt.js");
 
 var defaultConfigParameters = {
   defaultMaximumShownFeatureInfos: 100,
@@ -443,45 +444,6 @@ var Terria = function(options) {
    * @type {Boolean}
    */
   this.shouldStartSatelliteGuidance = false;
-
-  knockout.defineProperty(this, "pickedFeatures", {
-    get: () => {
-      return this._pickedFeatures;
-    },
-    set: value => {
-      this._pickedFeatures = value;
-      if (value.features) {
-        this.onFeaturePicked(value.features);
-      }
-    }
-  });
-
-  this.onFeaturePicked = features => {
-    if (
-      features &&
-      features.length > 0 &&
-      features[0].properties &&
-      features[0].properties["story-id"]
-    ) {
-      console.log("story-id:", features[0].properties["story-id"].valueOf());
-      //TODO handle multiple properties when multiple features are clicked at the same time (hotspot+line)
-      const hashProperties = {
-        share: features[0].properties["story-id"].valueOf(),
-        playStory: "1"
-      };
-      this.selectedFeature = features[0];
-      const temporaryInitSources = [];
-      interpretHash(
-        this,
-        hashProperties,
-        this.userProperties,
-        this.initSources,
-        temporaryInitSources
-      ).then(() => {
-        loadInitSources(this, temporaryInitSources);
-      });
-    }
-  };
 
   knockout.track(this, [
     "baseMaximumScreenSpaceError",
